+++ feature-utils/poly-look/src/react-components/index.js
@@ -3,7 +3,7 @@ import FilterChips from "./filterChips/filterChips.jsx";
 import { Tab, Tabs } from "./tabs.jsx";
 import { BlockLegend, LineLegend } from "./legends.jsx";
 import ErrorPopup from "./popUps/errorPopup/errorPopup.jsx";
-import { CardList, Card } from "./cardList/cardList.jsx";
+import { CardList, Card } from "./cards/cardList.jsx";
 
 export {
   Chip,
